Handle failures in synth call.
Because we don't rely on the provider doing the actual output to
device anymore, we don't need to worry about the service dissapearing
quietly in between method calls, and leaving us in a precarious state.

But, if the D-Bus Synthesize call fails and we don't handle it in our
queue, we are in trouble.