docs: fix missing annotation
GObject-Introspection now supports annotating the async/finish
functions, although the language bindings must support this to
allow automatic `await` support.

In the most current release the GIR compiler, and probably
gi-docgen, will emit warning if it can't match find a matching
pair of `some_async()`/`some_finish()`, as is the case with
`spiel_speaker_new()` and `spiel_speaker_new_finish()`.

The patch adds that annotation, but should be verified to pass the
CI on a GNOME 46 distribution before merge or it may result in an
error on older versions.